Refactor FeaturedSlider component to use a separate Slide component and improve carousel functionality with custom arrows. Add background image for slides.
